+++ serialization/src/jvmTestCommon/kotlin/nl/adaptivity/xmlutil/serialization/JvmSerializationTest.kt
@@ -20,8 +20,8 @@
 
 package nl.adaptivity.xmlutil.serialization
 
+import io.github.pdvrieze.xmlutil.testutil.assertXmlEquals
 import kotlinx.serialization.json.Json
-import nl.adaptivity.xml.serialization.assertXmlEquals
 import nl.adaptivity.xmlutil.XmlStreaming
 import org.junit.jupiter.api.Test
 import javax.xml.parsers.DocumentBuilderFactory